requirements.txt
fastapi
uvicorn
pillow
python-multipart

sentence-transformers
torch